added nlx-xrefs.ttl
This is the first pass a cleaned xrefs/altids. Pretty much in good
shape. oboInOwl:hasAlternativeId should go in the interlex existing ids,
we should hold off on the oboInOwl:hasDbXref ones because they are to
things that are relevant and related by are usually not vocabulary
sources.

One thing still to check for are that the same GBIF identifier can
appear on multiple classes, need to check about that.