ndc id postsubiculum, fix islands of calleja issue
There are a number of Olfactory tubercle Islets of Calleja cells that
were present in neurolex, the question is exactly what region these
neurons belong to.

For now I have ruled that they are part of UBERON:0023867
Islands of Calleja of olfactory tubercle, since those are referred to is
the Islets. Similar (or even the same type of) cells are probably also
located in other parts of the islands of calleja based on the references
available, however exactly which addtional regions needs more careful
review. The Isla magna of Calleja was probably a typo, though those
cells may also be present there, it does not appear to have been what
the original model intended.